refactor(server): absorb createVendoHandler route pair from @vendoai/next
