+++ packages/vendo-server/src/fetch-handler.ts
@@ -1,8 +1,10 @@
 /**
- * `createVendoFetchHandler()` — the framework-agnostic Web Fetch API
- * handler behind every Vendo framework adapter (`@vendoai/next` today).
- * Point any router at it — a Next.js catch-all, an Express/Hono bridge,
- * a raw `Bun.serve` — and it serves everything Vendo needs from one place.
+ * `createVendoFetchHandler()` — the framework-agnostic Web Fetch API core
+ * behind `createVendoHandler` (the `{ GET, POST }` route pair) and
+ * `toNodeHandler` (Express/`node:http`). The whole thing is surfaced
+ * publicly via `vendo/server`. Point any router at it — a Next.js
+ * catch-all, an Express/Hono bridge, a raw `Bun.serve` — and it serves
+ * everything Vendo needs from one place.
  *
  * Endpoints (relative to the route's mount, keyed on the LAST path segment
  * so any mount prefix works):
